Call `query_and_upload` asynchronously
https://github.com/project-koku/koku/issues/1408
