+++ koku/masu/util/upload.py
@@ -1,13 +1,7 @@
 """Upload utility functions."""
-import csv
 import logging
 
-from dateutil.rrule import DAILY, rrule
 from django.conf import settings
-from django.db import connection
-
-from api.dataexport.uploader import AwsS3Uploader
-from masu.util.common import NamedTemporaryGZip
 
 LOG = logging.getLogger(__name__)
 _DB_FETCH_BATCH_SIZE = 2000
@@ -51,65 +45,3 @@ def get_upload_path(
         )
     )
     return upload_path
-
-
-def query_and_upload_to_s3(
-    schema_name, provider_uuid, table_export_setting, date_range
-):
-    """
-    Query the database and upload the results to s3.
-
-    Args:
-        schema_name (str): Account schema name in which to execute the query.
-        provider_uuid (UUID): Provider UUID for filtering the query.
-        table_export_setting (TableExportSetting): Settings for the table export.
-        date_range (tuple): Pair of date objects of inclusive start and end dates.
-
-    """
-    LOG.info(
-        'query_and_upload_to_s3: schema %s provider_uuid %s table.output_name %s for %s',
-        schema_name,
-        provider_uuid,
-        table_export_setting.output_name,
-        date_range,
-    )
-    uploader = AwsS3Uploader(settings.S3_BUCKET_NAME)
-    start_date, end_date = date_range
-    iterate_daily = table_export_setting.iterate_daily
-    dates_to_iterate = rrule(
-        DAILY, dtstart=start_date, until=end_date if iterate_daily else start_date
-    )
-
-    for the_date in dates_to_iterate:
-        with NamedTemporaryGZip() as temp_file:
-            with connection.cursor() as cursor:
-                cursor.db.set_schema(schema_name)
-                upload_path = get_upload_path(
-                    schema_name,
-                    table_export_setting.provider,
-                    provider_uuid,
-                    the_date,
-                    table_export_setting.output_name,
-                    iterate_daily,
-                )
-                cursor.execute(
-                    table_export_setting.sql.format(schema=schema_name),
-                    {
-                        'start_date': the_date,
-                        'end_date': the_date if iterate_daily else end_date,
-                        'provider_uuid': provider_uuid,
-                    },
-                )
-                # Don't upload if result set is empty
-                if cursor.rowcount == 0:
-                    continue
-                writer = csv.writer(temp_file, quotechar='"', quoting=csv.QUOTE_MINIMAL)
-                writer.writerow([field.name for field in cursor.description])
-                while True:
-                    records = cursor.fetchmany(size=_DB_FETCH_BATCH_SIZE)
-                    if not records:
-                        break
-                    for row in records:
-                        writer.writerow(row)
-            temp_file.close()
-            uploader.upload_file(temp_file.name, upload_path)
